demoed version

+++ src/travel_recommender/model/abstracts/BaseGraph.java
@@ -10,7 +10,7 @@ public interface BaseGraph
 	
 	double get_edge_weight(BaseVertex source, BaseVertex sink);
 	Date get_edge_StartDate(BaseVertex source, BaseVertex sink);
-	double get_edge_Duration(BaseVertex source, BaseVertex sink);
+	long get_edge_Duration(BaseVertex source, BaseVertex sink);
 	double get_edge_Cost(BaseVertex source, BaseVertex sink);
 	Set<BaseVertex> get_adjacent_vertices(BaseVertex vertex);
 	Set<BaseVertex> get_precedent_vertices(BaseVertex vertex);
